% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horseshoe.R
\name{horseshoe}
\alias{horseshoe}
\title{A wrapper of horseshoe::horseshoe}
\usage{
horseshoe(x, y, ..., verbose = TRUE)
}
\arguments{
\item{x}{Matrix of covariates, dimension n*p.}

\item{y}{Response, a \eqn{n*1} vector.}

\item{...}{
  Arguments passed on to \code{\link[horseshoe:horseshoe]{horseshoe::horseshoe}}
  \describe{
    \item{\code{method.tau}}{Method for handling \eqn{\tau}. Select "truncatedCauchy" for full
Bayes with the Cauchy prior truncated to [1/p, 1], "halfCauchy" for full Bayes with
the half-Cauchy prior, or "fixed" to use a fixed value (an empirical Bayes estimate,
for example).}
    \item{\code{tau}}{Use this argument to pass the (estimated) value of \eqn{\tau} in case "fixed"
is selected for method.tau. Not necessary when method.tau is equal to"halfCauchy" or
"truncatedCauchy". The default (tau = 1) is not suitable for most purposes and should be replaced.}
    \item{\code{method.sigma}}{Select "Jeffreys" for full Bayes with Jeffrey's prior on the error
variance \eqn{\sigma^2}, or "fixed" to use a fixed value (an empirical Bayes
estimate, for example).}
    \item{\code{Sigma2}}{A fixed value for the error variance \eqn{\sigma^2}. Not necessary
when method.sigma is equal to "Jeffreys". Use this argument to pass the (estimated)
value of Sigma2 in case "fixed" is selected for method.sigma. The default (Sigma2 = 1)
is not suitable for most purposes and should be replaced.}
    \item{\code{burn}}{Number of burn-in MCMC samples. Default is 1000.}
    \item{\code{nmc}}{Number of posterior draws to be saved. Default is 5000.}
    \item{\code{thin}}{Thinning parameter of the chain. Default is 1 (no thinning).}
    \item{\code{alpha}}{Level for the credible intervals. For example, alpha = 0.05 results in
95\% credible intervals.}
  }}

\item{verbose}{\code{TRUE} (default) or \code{FALSE}.}
}
\description{
A wrapper of horseshoe::horseshoe
}
