% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horseshoe.R
\name{predict.horseshoe}
\alias{predict.horseshoe}
\title{Prediction method for horseshoe}
\usage{
\method{predict}{horseshoe}(object, newdata, posterior = c("mean", "median"), ...)
}
\arguments{
\item{object}{a model object for which prediction is desired.}

\item{newdata}{Matrix of covariates, dimension n*p.}

\item{posterior}{\code{"mean"} or \code{"median"}.}

\item{...}{additional arguments affecting the predictions produced.}
}
\description{
Prediction method for horseshoe
}
